requirements.txt
Django==1.9.6
python-social-auth==0.2.18
psycopg2==2.6.1
djangorestframework-3.3.3
markdown-2.6.6